feat: implemented CTA section to sign-in page. refactor: split code into more files and barrel index.
