+++ src/components/auth/Sections/index.jsx
@@ -1,0 +1,4 @@
+export { default as CallToAction } from './CTASection';
+export { default as Features } from './FeaturesSection';
+export { default as Hero } from './HeroSection';
+export { default as Stats } from './StatsSection';
