+++ src/components/auth/index.jsx
@@ -1,0 +1,13 @@
+import Main from './Main/Main';
+import Footer from './Footer/Footer';
+
+export const Layout = () => (
+
+    <>
+
+        <Main />
+        <Footer />
+
+    </>
+
+);
